#1060 Separate most components that are making API calls
