+++ frontend/components/card/CardDetails.vue
@@ -63,6 +63,5 @@ const eventId = typeof paramsEventId === "string" ? paramsEventId : undefined;
 
 const eventStore = useEventStore();
 await eventStore.fetchById(eventId);
-
 const { event } = eventStore;
 </script>
