+++ frontend/components/card/about/CardAboutEvent.vue
@@ -70,7 +70,6 @@ const eventId = typeof paramsEventId === "string" ? paramsEventId : undefined;
 
 const eventStore = useEventStore();
 await eventStore.fetchById(eventId);
-
 const { event } = eventStore;
 
 const description = ref();
