+++ frontend/components/card/about/CardAboutGroup.vue
@@ -92,7 +92,6 @@ const groupId = typeof paramsGroupId === "string" ? paramsGroupId : undefined;
 
 const groupStore = useGroupStore();
 await groupStore.fetchById(groupId);
-
 const { group } = groupStore;
 
 const description = ref();
