+++ frontend/components/card/get-involved/CardGetInvolvedEvent.vue
@@ -51,6 +51,5 @@ const eventId = typeof paramsEventId === "string" ? paramsEventId : undefined;
 
 const eventStore = useEventStore();
 await eventStore.fetchById(eventId);
-
 const { event } = eventStore;
 </script>
