+++ frontend/components/card/get-involved/CardGetInvolvedGroup.vue
@@ -48,6 +48,5 @@ const groupId = typeof paramsGroupId === "string" ? paramsGroupId : undefined;
 
 const groupStore = useGroupStore();
 await groupStore.fetchById(groupId);
-
 const { group } = groupStore;
 </script>
