+++ frontend/components/card/get-involved/CardGetInvolvedOrganization.vue
@@ -86,6 +86,5 @@ const orgId = typeof paramsOrgId === "string" ? paramsOrgId : undefined;
 
 const organizationStore = useOrganizationStore();
 await organizationStore.fetchById(orgId);
-
 const { organization } = organizationStore;
 </script>
