+++ frontend/components/feed/Feed.vue
@@ -20,7 +20,6 @@ const orgId = typeof paramsOrgId === "string" ? paramsOrgId : undefined;
 
 const organizationStore = useOrganizationStore();
 await organizationStore.fetchById(orgId);
-
 const { organization } = organizationStore;
 
 const feedItemNames = computed<string[]>(() => {
