+++ frontend/layouts/event/EventPage.vue
@@ -39,7 +39,6 @@ const eventId = typeof paramsEventId === "string" ? paramsEventId : undefined;
 
 const eventStore = useEventStore();
 await eventStore.fetchById(eventId);
-
 const { event } = eventStore;
 
 const { handleCloseModal: handleCloseModalUploadImages } =
