+++ frontend/layouts/group/GroupPage.vue
@@ -39,7 +39,6 @@ const groupId = typeof paramsGroupId === "string" ? paramsGroupId : undefined;
 
 const groupStore = useGroupStore();
 await groupStore.fetchById(groupId);
-
 const { group } = groupStore;
 
 const { handleCloseModal: handleCloseModalUploadImages } =
