+++ frontend/layouts/organization/OrganizationPage.vue
@@ -41,7 +41,6 @@ const orgId = typeof paramsOrgId === "string" ? paramsOrgId : undefined;
 
 const organizationStore = useOrganizationStore();
 await organizationStore.fetchById(orgId);
-
 const { organization } = organizationStore;
 
 const handleUploadComplete = async (fileUploadEntity: FileUploadEntity) => {
