+++ frontend/pages/events/[eventId]/resources.vue
@@ -6,8 +6,7 @@
         {{ event.name }}&nbsp;{{ $t("i18n._global.resources_lower") }}
       </Title>
     </Head>
-    <HeaderAppPage
-      :event="event"
+    <HeaderAppPageEvent
       :header="event.name + ' ' + $t('i18n._global.resources_lower')"
       :tagline="$t('i18n.pages.events.resources.tagline')"
       :underDevelopment="true"
@@ -24,7 +23,7 @@
           ariaLabel="i18n.pages._global.resources.new_resource_aria_label"
         />
       </div>
-    </HeaderAppPage>
+    </HeaderAppPageEvent>
     <div v-if="event.resources" class="space-y-3 py-4">
       <CardSearchResultResource
         v-for="(r, i) in event.resources"
