+++ frontend/pages/events/[eventId]/tasks.vue
@@ -6,8 +6,7 @@
         {{ event.name }}&nbsp;{{ $t("i18n.pages._global.tasks.tasks_lower") }}
       </Title>
     </Head>
-    <HeaderAppPage
-      :event="event"
+    <HeaderAppPageEvent
       :header="event.name + ' ' + $t('i18n.pages._global.tasks.tasks_lower')"
       :tagline="$t('i18n.pages._global.tasks.tasks_page_tagline')"
       :underDevelopment="true"
@@ -23,7 +22,7 @@
           ariaLabel="i18n.pages._global.tasks.new_task_aria_label"
         />
       </div>
-    </HeaderAppPage>
+    </HeaderAppPageEvent>
     <!-- <div v-if="event.tasks" class="space-y-3 py-4"></div>
     <EmptyState v-else pageType="tasks" :permission="false" /> -->
   </div>
