+++ frontend/pages/events/create.vue
@@ -95,7 +95,7 @@
             />
           </div>
           <div class="mx-14 mt-5 w-full">
-            <CardConnect pageType="event" />
+            <CardConnectEvent />
           </div>
           <div class="card-style mx-14 mt-5 w-full px-5 py-6">
             <label for="description" class="responsive-h3 block font-medium">
