+++ frontend/pages/organizations/[orgId]/affiliates.vue
@@ -8,8 +8,7 @@
         }}
       </Title>
     </Head>
-    <HeaderAppPage
-      :organization="organization"
+    <HeaderAppPageOrganization
       :header="
         organization.name +
         ' ' +
@@ -19,7 +18,7 @@
       :underDevelopment="true"
     >
       <div class="flex space-x-2 lg:space-x-3"></div>
-    </HeaderAppPage>
+    </HeaderAppPageOrganization>
     <!-- <div v-if="orgSupporters.length > 0" class="space-y-3 py-4">
       <CardSearchResultUser
         v-for="(u, i) in orgSupporters"
