+++ frontend/pages/organizations/[orgId]/groups/[groupId]/faq.vue
@@ -9,8 +9,7 @@
     <Head>
       <Title>{{ group.name }}&nbsp;{{ $t("i18n._global.faq") }}</Title>
     </Head>
-    <HeaderAppPage
-      :group="group"
+    <HeaderAppPageGroup
       :header="group.name + ' ' + $t('i18n._global.faq')"
       :tagline="$t('i18n.pages.organizations._global.faq_tagline')"
       :underDevelopment="true"
@@ -27,7 +26,7 @@
           ariaLabel="i18n.pages.organizations.groups._global.support_group_aria_label"
         /> -->
       </div>
-    </HeaderAppPage>
+    </HeaderAppPageGroup>
     <div v-if="group.faqEntries" class="py-4">
       <div v-for="f in group.faqEntries" class="mb-4">
         <CardFAQEntry :faqEntry="f" />
