+++ frontend/pages/organizations/[orgId]/resources.vue
@@ -6,8 +6,7 @@
         {{ organization.name }}&nbsp;{{ $t("i18n._global.resources_lower") }}
       </Title>
     </Head>
-    <HeaderAppPage
-      :organization="organization"
+    <HeaderAppPageOrganization
       :header="organization.name + ' ' + $t('i18n._global.resources_lower')"
       :tagline="$t('i18n.pages.organizations._global.resources_tagline')"
       :underDevelopment="true"
@@ -24,7 +23,7 @@
           ariaLabel="i18n.pages._global.resources.new_resource_aria_label"
         />
       </div>
-    </HeaderAppPage>
+    </HeaderAppPageOrganization>
     <div v-if="orgResources.length > 0" class="space-y-3 py-4">
       <CardSearchResultResource
         v-for="(r, i) in orgResources"
